token transfer amount & ibc token statistics

+++ internal/app/model/entity/ibc_token.go
@@ -16,7 +16,7 @@ type IBCToken struct {
 	Supply         string    `bson:"supply"`
 	TransferAmount string    `bson:"transfer_amount"`
 	TransferTxs    int64     `bson:"transfer_txs"`
-	ChainsInvolved int64     `bson:"chains_involved"`
+	ChainsInvolved int       `bson:"chains_involved"`
 }
 
 func (i IBCToken) CollectionName() string {
